🐛 코드 분할 및 상세 내용이 변경되지 않던 문제 해결
